feat(audio): create function to get audio from a clip
The function get_audio_from_clip simply takes a clip, gets it's audio and writes it as a .mp3 file. It returns the name of the file.

The function move_audio_to_correct_folder was also created. Since the function get_audio_from_clip creates mp3 files at the project root, there was the need for a single function to move these files to a specific folder. This is what move_audio_to_correct_folder does.

Some minor changes were also made, like a small change in the name of a function (which is now called transcribe_audio) and the path of the audio files generated by download_audio_from_youtube_url was changed from "audio" to "audios".